Move PendingRewardsData to shared-types, create a service function that controller calls
