+++ mercata/packages/shared-types/src/rewards.ts
@@ -1,0 +1,4 @@
+export interface PendingRewardsData {
+  pendingCata: string;
+  pendingCataFormatted: string;
+}
